% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-search.R
\name{assemble}
\alias{assemble}
\title{Assemble complete word search}
\usage{
assemble(grid, reference)
}
\arguments{
\item{grid}{A {ggplot2} object for the grid.}

\item{reference}{A {ggplot2} object for the lookup reference.}
}
\value{
An assembled {ggplot2} object with the grid and lookup reference.
}
\description{
`assemble()` creates a {ggplot2} object for previewing and assembling the final
visualization.
}
\examples{
# Wordsearch pieces
p1 <- render(demo_wordsearch$grid)
p2 <- render(demo_wordsearch$reference)

# Assembled wordsearch
assemble(p1, p2)
}
\seealso{
[render()]
}
