% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-search.R
\name{build_wordsearch}
\alias{build_wordsearch}
\title{Assemble complete word search}
\usage{
build_wordsearch(wordsearch, solution = TRUE)
}
\arguments{
\item{wordsearch}{A produced word search from `gen_word_search()` in list format, containing two elements:
'grid' and 'reference'. The function will replace all empty positions with randomly generated letters.}

\item{solution}{A logical (TRUE/FALSE) indicating if the solution should be rendered.}
}
\value{
An assembled {ggplot2} object with the grid, lookup reference, and solution
(if specified).
}
\description{
`build_wordsearch()` creates a {ggplot2} object of the assembled, final word search
}
\examples{
words <- c("Hi", "Hello", "Howdy")
wordsearch <- gen_word_search(words, "easy", "ready")

# Assembled wordsearch
build_wordsearch(wordsearch, TRUE)
}
\seealso{
[gen_word_search()]
}
