% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-search.R
\name{clean_up}
\alias{clean_up}
\title{Clean up word search matrix}
\usage{
clean_up(search)
}
\arguments{
\item{search}{A produced word search matrix.}
}
\value{
A cleaned, ready to render word search matrix.
}
\description{
`clean_up()` cleans a produced word search matrix, turning empty positions with
"-" into randomly generated characters.
}
\examples{
x <- matrix(
  c("-", "H", "-", "I"),
  nrow = 2,
  ncol = 2,
  byrow = TRUE
)
print(x)

y = clean_up(x)
print(y)
}
\seealso{
[gen_word_search()]
}
