% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_wordsearch}
\alias{demo_wordsearch}
\title{Demo word search}
\format{
## `demo_wordsearch`
A list of two data frame tables
}
\usage{
demo_wordsearch
}
\description{
A list of two data frame tables: 1) "grid" which contains the structure for
the word search itself and 2) "reference" which contains the structure for the
look up table.
}
\keyword{datasets}
