% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-search.R
\name{gen_word_search}
\alias{gen_word_search}
\title{Generate word search}
\usage{
gen_word_search(
  words,
  difficulty = c("easy", "hard", "mix"),
  type = c("ready", "raw")
)
}
\arguments{
\item{words}{A vector of character strings (words) required for the word search.}

\item{difficulty}{A character string specifying the difficulty, accepting
"easy", "hard", and "mix" values.}

\item{type}{A character string identifying the type of word search, accepting
"ready" and "raw" values. "ready" will create a fully populated word search while
"raw" will leave non-mapped positions with a "-" for validation.}
}
\value{
A list of two data frame table elements: 1) "grid", or the mapped words
ready for either render or validation, and 2) "reference", or the words to look
up in the grid.
}
\description{
`gen_word_search()` maps a vector of words onto a matrix.
}
\examples{
words <- c("Hi", "Hello", "Howdy")

# Creates a word search ready for rendering
wordsearch.ready <- gen_word_search(words, "easy", "ready")

# Creates a raw word search for validation
wordsearch.raw <- gen_word_search(words, "easy", "raw")
}
\seealso{
[ideal_dims(), clean_up(), get_mapping(), reorder_words()]
}
