% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-search.R
\name{get_mapping}
\alias{get_mapping}
\title{Make Tiles}
\usage{
get_mapping(dims, word, orientation)
}
\arguments{
\item{dims}{The dimensions of the matrix in number of rows and number of columns.}

\item{word}{A character string for mapping its placement in the word search matrix.}

\item{orientation}{A character string detailing its mapping direction, featuring
combinations of "top", "bottom", "left", and "right".}
}
\value{
A list of elements detailing the start position for mapping the word
and the its direction.
}
\description{
`get_mapping()` determines the possible positioning of the target word within
the word search matrix.
}
\examples{
words = c("Hi", "Hello", "Howdy")

reorder_words(words)
}
\seealso{
[gen_word_search()]
}
