% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-search.R
\name{ideal_dims}
\alias{ideal_dims}
\title{Ideal dimensions}
\usage{
ideal_dims(size)
}
\arguments{
\item{size}{An integer for the number of words intended for the search.}
}
\value{
A vector of length two for the dimensions of the reference matrix in
number of rows and number of columns.
}
\description{
`ideal_dims()` determines the proper dimensions for the reference matrix that
corresponds to the word search.
}
\examples{
ideal_dims(10)
}
\seealso{
[gen_word_search()]
}
