% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-search.R
\name{render}
\alias{render}
\title{Render word search}
\usage{
render(table)
}
\arguments{
\item{table}{A data frame table for plotting in a grid fashion.}
}
\value{
A {ggplot2} object.
}
\description{
`render()` creates a {ggplot2} object for previewing and assembling the final
visualization.
}
\examples{
# Rendered wordsearch
p <- render(demo_wordsearch$grid)
}
\seealso{
[gen_word_search(), assemble()]
}
