% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-search.R
\name{reorder_words}
\alias{reorder_words}
\title{Make Tiles}
\usage{
reorder_words(words)
}
\arguments{
\item{words}{A vector of character strings.}
}
\value{
A vector of words sorted by their character length.
}
\description{
`reorder_words()` reorders a vector of strings by their character length.
}
\examples{
words = c("Hi", "Hello", "Howdy")

reorder_words(words)
}
\seealso{
[gen_word_search()]
}
